Move event enums into separate file
